Change back to int for getFreeMemory.
Some apps are using this API and causing a runtime crash.

+++ core/java/android/os/Process.java
@@ -745,7 +745,7 @@ public class Process {
     public static final native void sendSignal(int pid, int signal);
     
     /** @hide */
-    public static final native long getFreeMemory();
+    public static final native int getFreeMemory();
     
     /** @hide */
     public static final native void readProcLines(String path,
